requirements.txt
scikit-learn
python-telegram-bot
python-dotenv
pytest
